requirements.txt
azure-functions==1.24.0
dnspython==2.8.0
MarkupSafe==3.0.3
pymongo==4.15.4
Werkzeug==3.1.3
